break up the visualisation script, but broken the websocket somehow
