+++ .env_template
@@ -38,10 +38,6 @@ MAX_RETRIES=3
 RETRY_DELAY=5
 API_CLIENT_TIMEOUT=30  # Timeout in seconds
 
-# Prompt and Topics
-# PROMPT="Your default prompt here"
-# TOPICS="Topic1,Topic2,Topic3"
-
 # Visualization Settings
 NODE_COLOR="0x1A0B31"
 EDGE_COLOR="0xff0000"
