+++ src/handlers/mod.rs
@@ -1,5 +1,9 @@
-// handlers/mod.rs
-
 pub mod file_handler;
 pub mod graph_handler;
 pub mod ragflow_handler;
+pub mod visualization_handler;  // New module for visualization settings
+
+pub use file_handler::*;
+pub use graph_handler::*;
+pub use ragflow_handler::*;
+pub use visualization_handler::*;  // Export visualization handler
